+++ apps/@examples/checklist-custom/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @examples/checklist-custom
 
+## 0.0.33
+
+### Patch Changes
+
+- @dopt/react-checklist@9.0.1
+- @dopt/react@4.1.1
+
 ## 0.0.32
 
 ### Patch Changes